split models to avoid circular import
